Upload Lab 2
This lab asks for the user to create two programs: one that calculates the mileage given the first and last odometer reading and another that converts Fahrenheit to Celsius. 

The purpose of the lab is to learn more about mathematical operations and user input in C. The lab also tests the nuances in C: For example, one must use floating point variables when dividing otherwise the expression will be calculated using integer division.