Upload Lab 3
Coding Problem 1: Calculate the monthly payment of a car when given the cost of the car, the down payment, the annual interest rate, and the number of payments. 

Coding Problem 2: Estimate the factorial of a given number.

The purpose of this lab is to use advanced concepts in mathematical computation in C. Concepts include precedence and the use of the math.h library which grants the pow() and sqrt() functions used to obtain the power and square root of a number respectively.